refactor : 함수 위치를 사용하는 파일로 옮기기
